Convert skeleton modifications to nodes.
