+++ editor/editor_node.cpp
@@ -183,7 +183,6 @@
 #include "editor/plugins/shader_file_editor_plugin.h"
 #include "editor/plugins/skeleton_2d_editor_plugin.h"
 #include "editor/plugins/skeleton_3d_editor_plugin.h"
-#include "editor/plugins/skeleton_ik_3d_editor_plugin.h"
 #include "editor/plugins/sprite_2d_editor_plugin.h"
 #include "editor/plugins/sprite_frames_editor_plugin.h"
 #include "editor/plugins/style_box_editor_plugin.h"
@@ -7169,7 +7168,6 @@ EditorNode::EditorNode() {
 	add_editor_plugin(memnew(ShaderEditorPlugin));
 	add_editor_plugin(memnew(ShaderFileEditorPlugin));
 	add_editor_plugin(memnew(Skeleton3DEditorPlugin));
-	add_editor_plugin(memnew(SkeletonIK3DEditorPlugin));
 	add_editor_plugin(memnew(SpriteFramesEditorPlugin));
 	add_editor_plugin(memnew(StyleBoxEditorPlugin));
 	add_editor_plugin(memnew(SubViewportPreviewEditorPlugin));
