+++ scene/2d/skeleton_modification_2d_ccdik.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  skeleton_modification_3d_ccdik.h                                     */
+/*  skeleton_modification_2d_ccdik.h                                     */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,87 +28,88 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#include "core/templates/local_vector.h"
-#include "scene/3d/skeleton_3d.h"
-#include "scene/resources/skeleton_modification_3d.h"
+#ifndef SKELETON_MODIFICATION_2D_CCDIK_H
+#define SKELETON_MODIFICATION_2D_CCDIK_H
 
-#ifndef SKELETON_MODIFICATION_3D_CCDIK_H
-#define SKELETON_MODIFICATION_3D_CCDIK_H
+#include "scene/2d/skeleton_2d.h"
+#include "scene/2d/skeleton_modification_2d.h"
 
-class SkeletonModification3DCCDIK : public SkeletonModification3D {
-	GDCLASS(SkeletonModification3DCCDIK, SkeletonModification3D);
+///////////////////////////////////////
+// SkeletonModification2DCCDIK
+///////////////////////////////////////
+
+class SkeletonModification2DCCDIK : public SkeletonModification2D {
+	GDCLASS(SkeletonModification2DCCDIK, SkeletonModification2D);
 
 private:
-	enum CCDIK_Axes {
-		AXIS_X,
-		AXIS_Y,
-		AXIS_Z
-	};
-
-	struct CCDIK_Joint_Data {
-		String bone_name = "";
+	struct CCDIK_Joint_Data2D {
 		int bone_idx = -1;
-		int ccdik_axis = 0;
+		NodePath bone2d_node;
+		ObjectID bone2d_node_cache;
+		bool rotate_from_joint = false;
 
 		bool enable_constraint = false;
-		real_t constraint_angle_min = 0;
-		real_t constraint_angle_max = (2.0 * Math_PI);
-		bool constraint_angles_invert = false;
+		float constraint_angle_min = 0;
+		float constraint_angle_max = (2.0 * Math_PI);
+		bool constraint_angle_invert = false;
+		bool constraint_in_localspace = true;
+
+		bool editor_draw_gizmo = true;
 	};
 
-	LocalVector<CCDIK_Joint_Data> ccdik_data_chain;
+	Vector<CCDIK_Joint_Data2D> ccdik_data_chain;
+
 	NodePath target_node;
 	ObjectID target_node_cache;
+	void update_target_cache();
 
 	NodePath tip_node;
 	ObjectID tip_node_cache;
-
-	bool use_high_quality_solve = true;
-
-	void update_target_cache();
 	void update_tip_cache();
 
-	void _execute_ccdik_joint(int p_joint_idx, Node3D *p_target, Node3D *p_tip);
+	void ccdik_joint_update_bone2d_cache(int p_joint_idx);
+	void _execute_ccdik_joint(int p_joint_idx, Node2D *p_target, Node2D *p_tip);
 
 protected:
 	static void _bind_methods();
-	bool _get(const StringName &p_path, Variant &r_ret) const;
 	bool _set(const StringName &p_path, const Variant &p_value);
+	bool _get(const StringName &p_path, Variant &r_ret) const;
 	void _get_property_list(List<PropertyInfo> *p_list) const;
 
 public:
-	virtual void _execute(real_t p_delta) override;
-	virtual void _setup_modification(SkeletonModificationStack3D *p_stack) override;
+	void _notification(int p_what);
+	void _draw_editor_gizmo() override;
 
 	void set_target_node(const NodePath &p_target_node);
 	NodePath get_target_node() const;
-
 	void set_tip_node(const NodePath &p_tip_node);
 	NodePath get_tip_node() const;
-
-	void set_use_high_quality_solve(bool p_solve);
-	bool get_use_high_quality_solve() const;
-
-	String get_ccdik_joint_bone_name(int p_joint_idx) const;
-	void set_ccdik_joint_bone_name(int p_joint_idx, String p_bone_name);
-	int get_ccdik_joint_bone_index(int p_joint_idx) const;
-	void set_ccdik_joint_bone_index(int p_joint_idx, int p_bone_idx);
-	int get_ccdik_joint_ccdik_axis(int p_joint_idx) const;
-	void set_ccdik_joint_ccdik_axis(int p_joint_idx, int p_axis);
-	bool get_ccdik_joint_enable_constraint(int p_joint_idx) const;
-	void set_ccdik_joint_enable_constraint(int p_joint_idx, bool p_enable);
-	real_t get_ccdik_joint_constraint_angle_min(int p_joint_idx) const;
-	void set_ccdik_joint_constraint_angle_min(int p_joint_idx, real_t p_angle_min);
-	real_t get_ccdik_joint_constraint_angle_max(int p_joint_idx) const;
-	void set_ccdik_joint_constraint_angle_max(int p_joint_idx, real_t p_angle_max);
-	bool get_ccdik_joint_constraint_invert(int p_joint_idx) const;
-	void set_ccdik_joint_constraint_invert(int p_joint_idx, bool p_invert);
 
 	int get_ccdik_data_chain_length();
 	void set_ccdik_data_chain_length(int p_new_length);
 
-	SkeletonModification3DCCDIK();
-	~SkeletonModification3DCCDIK();
+	void set_ccdik_joint_bone2d_node(int p_joint_idx, const NodePath &p_target_node);
+	NodePath get_ccdik_joint_bone2d_node(int p_joint_idx) const;
+	void set_ccdik_joint_bone_index(int p_joint_idx, int p_bone_idx);
+	int get_ccdik_joint_bone_index(int p_joint_idx) const;
+
+	void set_ccdik_joint_rotate_from_joint(int p_joint_idx, bool p_rotate_from_joint);
+	bool get_ccdik_joint_rotate_from_joint(int p_joint_idx) const;
+	void set_ccdik_joint_enable_constraint(int p_joint_idx, bool p_constraint);
+	bool get_ccdik_joint_enable_constraint(int p_joint_idx) const;
+	void set_ccdik_joint_constraint_angle_min(int p_joint_idx, float p_angle_min);
+	float get_ccdik_joint_constraint_angle_min(int p_joint_idx) const;
+	void set_ccdik_joint_constraint_angle_max(int p_joint_idx, float p_angle_max);
+	float get_ccdik_joint_constraint_angle_max(int p_joint_idx) const;
+	void set_ccdik_joint_constraint_angle_invert(int p_joint_idx, bool p_invert);
+	bool get_ccdik_joint_constraint_angle_invert(int p_joint_idx) const;
+	void set_ccdik_joint_constraint_in_localspace(int p_joint_idx, bool p_constraint_in_localspace);
+	bool get_ccdik_joint_constraint_in_localspace(int p_joint_idx) const;
+	void set_ccdik_joint_editor_draw_gizmo(int p_joint_idx, bool p_draw_gizmo);
+	bool get_ccdik_joint_editor_draw_gizmo(int p_joint_idx) const;
+
+	SkeletonModification2DCCDIK();
+	~SkeletonModification2DCCDIK();
 };
 
-#endif // SKELETON_MODIFICATION_3D_CCDIK_H
+#endif // SKELETON_MODIFICATION_2D_CCDIK_H
