+++ scene/2d/skeleton_modification_2d_fabrik.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  skeleton_modification_2d_physicalbones.h                             */
+/*  skeleton_modification_2d_fabrik.h                                    */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,55 +28,80 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef SKELETON_MODIFICATION_2D_PHYSICALBONES_H
-#define SKELETON_MODIFICATION_2D_PHYSICALBONES_H
+#ifndef SKELETON_MODIFICATION_2D_FABRIK_H
+#define SKELETON_MODIFICATION_2D_FABRIK_H
 
 #include "scene/2d/skeleton_2d.h"
-#include "scene/resources/skeleton_modification_2d.h"
+#include "scene/2d/skeleton_modification_2d.h"
 
 ///////////////////////////////////////
-// SkeletonModification2DJIGGLE
+// SkeletonModification2DFABRIK
 ///////////////////////////////////////
 
-class SkeletonModification2DPhysicalBones : public SkeletonModification2D {
-	GDCLASS(SkeletonModification2DPhysicalBones, SkeletonModification2D);
+class SkeletonModification2DFABRIK : public SkeletonModification2D {
+	GDCLASS(SkeletonModification2DFABRIK, SkeletonModification2D);
 
 private:
-	struct PhysicalBone_Data2D {
-		NodePath physical_bone_node;
-		ObjectID physical_bone_node_cache;
+	struct FABRIK_Joint_Data2D {
+		int bone_idx = -1;
+		NodePath bone2d_node;
+		ObjectID bone2d_node_cache;
+
+		Vector2 magnet_position = Vector2(0, 0);
+		bool use_target_rotation = false;
+
+		bool editor_draw_gizmo = true;
 	};
-	Vector<PhysicalBone_Data2D> physical_bone_chain;
 
-	void _physical_bone_update_cache(int p_joint_idx);
+	Vector<FABRIK_Joint_Data2D> fabrik_data_chain;
 
-	bool _simulation_state_dirty = false;
-	TypedArray<StringName> _simulation_state_dirty_names;
-	bool _simulation_state_dirty_process = false;
-	void _update_simulation_state();
+	// Unlike in 3D, we need a vector of Transform2D objects to perform FABRIK.
+	// This is because FABRIK (unlike CCDIK) needs to operate on transforms that are NOT
+	// affected by each other, making the transforms stored in Bone2D unusable, as well as those in Skeleton2D.
+	// For this reason, this modification stores a vector of Transform2Ds used for the calculations, which are then applied at the end.
+	Vector<Transform2D> fabrik_transform_chain;
+
+	NodePath target_node;
+	ObjectID target_node_cache;
+	void update_target_cache();
+
+	float chain_tolarance = 0.01;
+	int chain_max_iterations = 10;
+	int chain_iterations = 0;
+	Transform2D target_global_pose = Transform2D();
+	Transform2D origin_global_pose = Transform2D();
+
+	void fabrik_joint_update_bone2d_cache(int p_joint_idx);
+	void chain_backwards();
+	void chain_forwards();
 
 protected:
 	static void _bind_methods();
-	bool _get(const StringName &p_path, Variant &r_ret) const;
 	bool _set(const StringName &p_path, const Variant &p_value);
+	bool _get(const StringName &p_path, Variant &r_ret) const;
 	void _get_property_list(List<PropertyInfo> *p_list) const;
 
 public:
-	void _execute(float p_delta) override;
-	void _setup_modification(SkeletonModificationStack2D *p_stack) override;
+	void _notification(int p_what);
 
-	int get_physical_bone_chain_length();
-	void set_physical_bone_chain_length(int p_new_length);
+	void set_target_node(const NodePath &p_target_node);
+	NodePath get_target_node() const;
 
-	void set_physical_bone_node(int p_joint_idx, const NodePath &p_path);
-	NodePath get_physical_bone_node(int p_joint_idx) const;
+	int get_fabrik_data_chain_length();
+	void set_fabrik_data_chain_length(int p_new_length);
 
-	void fetch_physical_bones();
-	void start_simulation(const TypedArray<StringName> &p_bones);
-	void stop_simulation(const TypedArray<StringName> &p_bones);
+	void set_fabrik_joint_bone2d_node(int p_joint_idx, const NodePath &p_target_node);
+	NodePath get_fabrik_joint_bone2d_node(int p_joint_idx) const;
+	void set_fabrik_joint_bone_index(int p_joint_idx, int p_bone_idx);
+	int get_fabrik_joint_bone_index(int p_joint_idx) const;
 
-	SkeletonModification2DPhysicalBones();
-	~SkeletonModification2DPhysicalBones();
+	void set_fabrik_joint_magnet_position(int p_joint_idx, Vector2 p_magnet_position);
+	Vector2 get_fabrik_joint_magnet_position(int p_joint_idx) const;
+	void set_fabrik_joint_use_target_rotation(int p_joint_idx, bool p_use_target_rotation);
+	bool get_fabrik_joint_use_target_rotation(int p_joint_idx) const;
+
+	SkeletonModification2DFABRIK();
+	~SkeletonModification2DFABRIK();
 };
 
-#endif // SKELETON_MODIFICATION_2D_PHYSICALBONES_H
+#endif // SKELETON_MODIFICATION_2D_FABRIK_H
