+++ scene/2d/skeleton_modification_2d_jiggle.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  skeleton_modification_3d_jiggle.h                                    */
+/*  skeleton_modification_2d_jiggle.h                                    */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,80 +28,82 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#include "core/templates/local_vector.h"
-#include "scene/3d/skeleton_3d.h"
-#include "scene/resources/skeleton_modification_3d.h"
+#ifndef SKELETON_MODIFICATION_2D_JIGGLE_H
+#define SKELETON_MODIFICATION_2D_JIGGLE_H
 
-#ifndef SKELETON_MODIFICATION_3D_JIGGLE_H
-#define SKELETON_MODIFICATION_3D_JIGGLE_H
+#include "scene/2d/skeleton_2d.h"
+#include "scene/2d/skeleton_modification_2d.h"
 
-class SkeletonModification3DJiggle : public SkeletonModification3D {
-	GDCLASS(SkeletonModification3DJiggle, SkeletonModification3D);
+///////////////////////////////////////
+// SkeletonModification2DJIGGLE
+///////////////////////////////////////
+
+class SkeletonModification2DJiggle : public SkeletonModification2D {
+	GDCLASS(SkeletonModification2DJiggle, SkeletonModification2D);
 
 private:
-	struct Jiggle_Joint_Data {
-		String bone_name = "";
+	struct Jiggle_Joint_Data2D {
 		int bone_idx = -1;
+		NodePath bone2d_node;
+		ObjectID bone2d_node_cache;
 
 		bool override_defaults = false;
-		real_t stiffness = 3;
-		real_t mass = 0.75;
-		real_t damping = 0.75;
+		float stiffness = 3;
+		float mass = 0.75;
+		float damping = 0.75;
 		bool use_gravity = false;
-		Vector3 gravity = Vector3(0, -6.0, 0);
-		real_t roll = 0;
+		Vector2 gravity = Vector2(0, 6.0);
 
-		Vector3 cached_rotation = Vector3(0, 0, 0);
-		Vector3 force = Vector3(0, 0, 0);
-		Vector3 acceleration = Vector3(0, 0, 0);
-		Vector3 velocity = Vector3(0, 0, 0);
-		Vector3 last_position = Vector3(0, 0, 0);
-		Vector3 dynamic_position = Vector3(0, 0, 0);
+		Vector2 force = Vector2(0, 0);
+		Vector2 acceleration = Vector2(0, 0);
+		Vector2 velocity = Vector2(0, 0);
+		Vector2 last_position = Vector2(0, 0);
+		Vector2 dynamic_position = Vector2(0, 0);
 
-		Vector3 last_noncollision_position = Vector3(0, 0, 0);
+		Vector2 last_noncollision_position = Vector2(0, 0);
 	};
+
+	Vector<Jiggle_Joint_Data2D> jiggle_data_chain;
 
 	NodePath target_node;
 	ObjectID target_node_cache;
-	LocalVector<Jiggle_Joint_Data> jiggle_data_chain;
+	void update_target_cache();
 
-	real_t stiffness = 3;
-	real_t mass = 0.75;
-	real_t damping = 0.75;
+	float stiffness = 3;
+	float mass = 0.75;
+	float damping = 0.75;
 	bool use_gravity = false;
-	Vector3 gravity = Vector3(0, -6.0, 0);
+	Vector2 gravity = Vector2(0, 6);
 
 	bool use_colliders = false;
 	uint32_t collision_mask = 1;
 
-	void update_cache();
-	void _execute_jiggle_joint(int p_joint_idx, Node3D *p_target, real_t p_delta);
+	void jiggle_joint_update_bone2d_cache(int p_joint_idx);
+	void _execute_jiggle_joint(int p_joint_idx, Node2D *p_target, float p_delta);
 	void _update_jiggle_joint_data();
 
 protected:
 	static void _bind_methods();
-	bool _get(const StringName &p_path, Variant &r_ret) const;
 	bool _set(const StringName &p_path, const Variant &p_value);
+	bool _get(const StringName &p_path, Variant &r_ret) const;
 	void _get_property_list(List<PropertyInfo> *p_list) const;
 
 public:
-	virtual void _execute(real_t p_delta) override;
-	virtual void _setup_modification(SkeletonModificationStack3D *p_stack) override;
+	void _notification(int p_what);
 
 	void set_target_node(const NodePath &p_target_node);
 	NodePath get_target_node() const;
 
-	void set_stiffness(real_t p_stiffness);
-	real_t get_stiffness() const;
-	void set_mass(real_t p_mass);
-	real_t get_mass() const;
-	void set_damping(real_t p_damping);
-	real_t get_damping() const;
-
+	void set_stiffness(float p_stiffness);
+	float get_stiffness() const;
+	void set_mass(float p_mass);
+	float get_mass() const;
+	void set_damping(float p_damping);
+	float get_damping() const;
 	void set_use_gravity(bool p_use_gravity);
 	bool get_use_gravity() const;
-	void set_gravity(Vector3 p_gravity);
-	Vector3 get_gravity() const;
+	void set_gravity(Vector2 p_gravity);
+	Vector2 get_gravity() const;
 
 	void set_use_colliders(bool p_use_colliders);
 	bool get_use_colliders() const;
@@ -111,28 +113,26 @@ public:
 	int get_jiggle_data_chain_length();
 	void set_jiggle_data_chain_length(int p_new_length);
 
-	void set_jiggle_joint_bone_name(int p_joint_idx, String p_name);
-	String get_jiggle_joint_bone_name(int p_joint_idx) const;
-	void set_jiggle_joint_bone_index(int p_joint_idx, int p_idx);
+	void set_jiggle_joint_bone2d_node(int p_joint_idx, const NodePath &p_target_node);
+	NodePath get_jiggle_joint_bone2d_node(int p_joint_idx) const;
+	void set_jiggle_joint_bone_index(int p_joint_idx, int p_bone_idx);
 	int get_jiggle_joint_bone_index(int p_joint_idx) const;
 
 	void set_jiggle_joint_override(int p_joint_idx, bool p_override);
 	bool get_jiggle_joint_override(int p_joint_idx) const;
-	void set_jiggle_joint_stiffness(int p_joint_idx, real_t p_stiffness);
-	real_t get_jiggle_joint_stiffness(int p_joint_idx) const;
-	void set_jiggle_joint_mass(int p_joint_idx, real_t p_mass);
-	real_t get_jiggle_joint_mass(int p_joint_idx) const;
-	void set_jiggle_joint_damping(int p_joint_idx, real_t p_damping);
-	real_t get_jiggle_joint_damping(int p_joint_idx) const;
+	void set_jiggle_joint_stiffness(int p_joint_idx, float p_stiffness);
+	float get_jiggle_joint_stiffness(int p_joint_idx) const;
+	void set_jiggle_joint_mass(int p_joint_idx, float p_mass);
+	float get_jiggle_joint_mass(int p_joint_idx) const;
+	void set_jiggle_joint_damping(int p_joint_idx, float p_damping);
+	float get_jiggle_joint_damping(int p_joint_idx) const;
 	void set_jiggle_joint_use_gravity(int p_joint_idx, bool p_use_gravity);
 	bool get_jiggle_joint_use_gravity(int p_joint_idx) const;
-	void set_jiggle_joint_gravity(int p_joint_idx, Vector3 p_gravity);
-	Vector3 get_jiggle_joint_gravity(int p_joint_idx) const;
-	void set_jiggle_joint_roll(int p_joint_idx, real_t p_roll);
-	real_t get_jiggle_joint_roll(int p_joint_idx) const;
+	void set_jiggle_joint_gravity(int p_joint_idx, Vector2 p_gravity);
+	Vector2 get_jiggle_joint_gravity(int p_joint_idx) const;
 
-	SkeletonModification3DJiggle();
-	~SkeletonModification3DJiggle();
+	SkeletonModification2DJiggle();
+	~SkeletonModification2DJiggle();
 };
 
-#endif // SKELETON_MODIFICATION_3D_JIGGLE_H
+#endif // SKELETON_MODIFICATION_2D_JIGGLE_H
