+++ scene/2d/skeleton_modification_2d_twoboneik.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  skeleton_modification_2d_fabrik.h                                    */
+/*  skeleton_modification_2d_twoboneik.h                                 */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,81 +28,79 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef SKELETON_MODIFICATION_2D_FABRIK_H
-#define SKELETON_MODIFICATION_2D_FABRIK_H
+#ifndef SKELETON_MODIFICATION_2D_TWOBONEIK_H
+#define SKELETON_MODIFICATION_2D_TWOBONEIK_H
 
 #include "scene/2d/skeleton_2d.h"
-#include "scene/resources/skeleton_modification_2d.h"
+#include "scene/2d/skeleton_modification_2d.h"
 
 ///////////////////////////////////////
-// SkeletonModification2DFABRIK
+// SkeletonModification2DJIGGLE
 ///////////////////////////////////////
 
-class SkeletonModification2DFABRIK : public SkeletonModification2D {
-	GDCLASS(SkeletonModification2DFABRIK, SkeletonModification2D);
+class SkeletonModification2DTwoBoneIK : public SkeletonModification2D {
+	GDCLASS(SkeletonModification2DTwoBoneIK, SkeletonModification2D);
 
 private:
-	struct FABRIK_Joint_Data2D {
-		int bone_idx = -1;
-		NodePath bone2d_node;
-		ObjectID bone2d_node_cache;
-
-		Vector2 magnet_position = Vector2(0, 0);
-		bool use_target_rotation = false;
-
-		bool editor_draw_gizmo = true;
-	};
-
-	Vector<FABRIK_Joint_Data2D> fabrik_data_chain;
-
-	// Unlike in 3D, we need a vector of Transform2D objects to perform FABRIK.
-	// This is because FABRIK (unlike CCDIK) needs to operate on transforms that are NOT
-	// affected by each other, making the transforms stored in Bone2D unusable, as well as those in Skeleton2D.
-	// For this reason, this modification stores a vector of Transform2Ds used for the calculations, which are then applied at the end.
-	Vector<Transform2D> fabrik_transform_chain;
-
 	NodePath target_node;
 	ObjectID target_node_cache;
+	float target_minimum_distance = 0;
+	float target_maximum_distance = 0;
+	bool flip_bend_direction = false;
+
+	NodePath joint_one_bone2d_node;
+	ObjectID joint_one_bone2d_node_cache;
+	int joint_one_bone_idx = -1;
+
+	NodePath joint_two_bone2d_node;
+	ObjectID joint_two_bone2d_node_cache;
+	int joint_two_bone_idx = -1;
+
+#ifdef TOOLS_ENABLED
+	bool editor_draw_min_max = false;
+#endif // TOOLS_ENABLED
+
 	void update_target_cache();
-
-	float chain_tolarance = 0.01;
-	int chain_max_iterations = 10;
-	int chain_iterations = 0;
-	Transform2D target_global_pose = Transform2D();
-	Transform2D origin_global_pose = Transform2D();
-
-	void fabrik_joint_update_bone2d_cache(int p_joint_idx);
-	void chain_backwards();
-	void chain_forwards();
+	void update_joint_one_bone2d_cache();
+	void update_joint_two_bone2d_cache();
 
 protected:
 	static void _bind_methods();
-	bool _set(const StringName &p_path, const Variant &p_value);
 	bool _get(const StringName &p_path, Variant &r_ret) const;
+	bool _set(const StringName &p_path, const Variant &p_value);
 	void _get_property_list(List<PropertyInfo> *p_list) const;
 
 public:
-	void _execute(float p_delta) override;
-	void _setup_modification(SkeletonModificationStack2D *p_stack) override;
+	void _notification(int p_what);
+	void _draw_editor_gizmo() override;
 
 	void set_target_node(const NodePath &p_target_node);
 	NodePath get_target_node() const;
 
-	int get_fabrik_data_chain_length();
-	void set_fabrik_data_chain_length(int p_new_length);
+	void set_target_minimum_distance(float p_minimum_distance);
+	float get_target_minimum_distance() const;
+	void set_target_maximum_distance(float p_maximum_distance);
+	float get_target_maximum_distance() const;
+	void set_flip_bend_direction(bool p_flip_direction);
+	bool get_flip_bend_direction() const;
 
-	void set_fabrik_joint_bone2d_node(int p_joint_idx, const NodePath &p_target_node);
-	NodePath get_fabrik_joint_bone2d_node(int p_joint_idx) const;
-	void set_fabrik_joint_bone_index(int p_joint_idx, int p_bone_idx);
-	int get_fabrik_joint_bone_index(int p_joint_idx) const;
+	void set_joint_one_bone2d_node(const NodePath &p_node);
+	NodePath get_joint_one_bone2d_node() const;
+	void set_joint_one_bone_idx(int p_bone_idx);
+	int get_joint_one_bone_idx() const;
 
-	void set_fabrik_joint_magnet_position(int p_joint_idx, Vector2 p_magnet_position);
-	Vector2 get_fabrik_joint_magnet_position(int p_joint_idx) const;
-	void set_fabrik_joint_use_target_rotation(int p_joint_idx, bool p_use_target_rotation);
-	bool get_fabrik_joint_use_target_rotation(int p_joint_idx) const;
+	void set_joint_two_bone2d_node(const NodePath &p_node);
+	NodePath get_joint_two_bone2d_node() const;
+	void set_joint_two_bone_idx(int p_bone_idx);
+	int get_joint_two_bone_idx() const;
 
-	SkeletonModification2DFABRIK();
-	~SkeletonModification2DFABRIK();
+#ifdef TOOLS_ENABLED
+	void set_editor_draw_min_max(bool p_draw);
+	bool get_editor_draw_min_max() const;
+#endif // TOOLS_ENABLED
+
+	SkeletonModification2DTwoBoneIK();
+	~SkeletonModification2DTwoBoneIK();
 };
 
-#endif // SKELETON_MODIFICATION_2D_FABRIK_H
+#endif // SKELETON_MODIFICATION_2D_TWOBONEIK_H
