+++ scene/3d/skeleton_modification_3d_twoboneik.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  skeleton_modification_2d_stackholder.h                               */
+/*  skeleton_modification_3d_twoboneik.h                                 */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,37 +28,89 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef SKELETON_MODIFICATION_2D_STACKHOLDER_H
-#define SKELETON_MODIFICATION_2D_STACKHOLDER_H
+#include "scene/3d/skeleton_3d.h"
+#include "scene/3d/skeleton_modification_3d.h"
 
-#include "scene/2d/skeleton_2d.h"
-#include "scene/resources/skeleton_modification_2d.h"
+#ifndef SKELETON_MODIFICATION_3D_TWOBONEIK_H
+#define SKELETON_MODIFICATION_3D_TWOBONEIK_H
 
-///////////////////////////////////////
-// SkeletonModification2DJIGGLE
-///////////////////////////////////////
+class SkeletonModification3DTwoBoneIK : public SkeletonModification3D {
+	GDCLASS(SkeletonModification3DTwoBoneIK, SkeletonModification3D);
 
-class SkeletonModification2DStackHolder : public SkeletonModification2D {
-	GDCLASS(SkeletonModification2DStackHolder, SkeletonModification2D);
+private:
+	NodePath target_node;
+	ObjectID target_node_cache;
+
+	bool use_tip_node = false;
+	NodePath tip_node;
+	ObjectID tip_node_cache;
+
+	bool use_pole_node = false;
+	NodePath pole_node;
+	ObjectID pole_node_cache;
+
+	String joint_one_bone_name = "";
+	int joint_one_bone_idx = -1;
+	String joint_two_bone_name = "";
+	int joint_two_bone_idx = -1;
+
+	bool auto_calculate_joint_length = false;
+	real_t joint_one_length = -1;
+	real_t joint_two_length = -1;
+
+	real_t joint_one_roll = 0;
+	real_t joint_two_roll = 0;
+
+	void update_cache_target();
+	void update_cache_tip();
+	void update_cache_pole();
 
 protected:
+	void _notification(int32_t p_what);
 	static void _bind_methods();
 	bool _get(const StringName &p_path, Variant &r_ret) const;
 	bool _set(const StringName &p_path, const Variant &p_value);
 	void _get_property_list(List<PropertyInfo> *p_list) const;
 
 public:
-	Ref<SkeletonModificationStack2D> held_modification_stack;
+	void set_target_node(const NodePath &p_target_node);
+	NodePath get_target_node() const;
 
-	void _execute(float p_delta) override;
-	void _setup_modification(SkeletonModificationStack2D *p_stack) override;
-	void _draw_editor_gizmo() override;
+	void set_use_tip_node(const bool p_use_tip_node);
+	bool get_use_tip_node() const;
+	void set_tip_node(const NodePath &p_tip_node);
+	NodePath get_tip_node() const;
 
-	void set_held_modification_stack(Ref<SkeletonModificationStack2D> p_held_stack);
-	Ref<SkeletonModificationStack2D> get_held_modification_stack() const;
+	void set_use_pole_node(const bool p_use_pole_node);
+	bool get_use_pole_node() const;
+	void set_pole_node(const NodePath &p_pole_node);
+	NodePath get_pole_node() const;
 
-	SkeletonModification2DStackHolder();
-	~SkeletonModification2DStackHolder();
+	void set_auto_calculate_joint_length(bool p_calculate);
+	bool get_auto_calculate_joint_length() const;
+	void calculate_joint_lengths();
+
+	void set_joint_one_bone_name(String p_bone_name);
+	String get_joint_one_bone_name() const;
+	void set_joint_one_bone_idx(int p_bone_idx);
+	int get_joint_one_bone_idx() const;
+	void set_joint_one_length(real_t p_length);
+	real_t get_joint_one_length() const;
+
+	void set_joint_two_bone_name(String p_bone_name);
+	String get_joint_two_bone_name() const;
+	void set_joint_two_bone_idx(int p_bone_idx);
+	int get_joint_two_bone_idx() const;
+	void set_joint_two_length(real_t p_length);
+	real_t get_joint_two_length() const;
+
+	void set_joint_one_roll(real_t p_roll);
+	real_t get_joint_one_roll() const;
+	void set_joint_two_roll(real_t p_roll);
+	real_t get_joint_two_roll() const;
+
+	SkeletonModification3DTwoBoneIK();
+	~SkeletonModification3DTwoBoneIK();
 };
 
-#endif // SKELETON_MODIFICATION_2D_STACKHOLDER_H
+#endif // SKELETON_MODIFICATION_3D_TWOBONEIK_H
